+++ integration/setup/setup.go
@@ -16,8 +16,8 @@ import (
 	"github.com/giantswarm/microerror"
 	"github.com/spf13/afero"
 
-	"github.com/giantswarm/app-operator/v2/integration/key"
-	"github.com/giantswarm/app-operator/v2/pkg/project"
+	"github.com/giantswarm/app-operator/v3/integration/key"
+	"github.com/giantswarm/app-operator/v3/pkg/project"
 )
 
 func Setup(m *testing.M, config Config) {