+++ service/controller/app/resource/chartoperator/delete.go
@@ -6,7 +6,7 @@ import (
 	"github.com/giantswarm/app/v4/pkg/key"
 	"github.com/giantswarm/microerror"
 
-	"github.com/giantswarm/app-operator/v2/service/controller/app/controllercontext"
+	"github.com/giantswarm/app-operator/v3/service/controller/app/controllercontext"
 )
 
 func (r Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {