+++ service/controller/app/resource/clients/resource.go
@@ -15,7 +15,7 @@ import (
 	"github.com/spf13/afero"
 	"k8s.io/client-go/rest"
 
-	"github.com/giantswarm/app-operator/v2/service/controller/app/controllercontext"
+	"github.com/giantswarm/app-operator/v3/service/controller/app/controllercontext"
 )
 
 const (