+++ CHANGELOG.md
@@ -7,8 +7,18 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [3.0.0] - 2021-01-05
+
+### Changed
+
+- Enable mutating and validating webhooks in app-admission-controller for
+tenant app CRs.
+
 ### Added
 
+- Make resync period configurable for use in integration tests.
+- Pause App CR reconciliation when it has
+  `app-operator.giantswarm.io/paused=true` annotation.
 - Print difference between the current chart and desired chart.
 
 ## [2.8.0] - 2020-12-15
@@ -151,7 +161,7 @@ app CRs. A message is added to the app CR status for the user.
 
 ### Removed
 
-- Don't wait for chart-operator pod since chart CRD is created by the chartcrd resource. 
+- Don't wait for chart-operator pod since chart CRD is created by the chartcrd resource.
 
 ## [2.0.0] - 2020-08-13
 
@@ -262,7 +272,8 @@ from Helm 2 to Helm 3.
 
 - Flattening operator release structure.
 
-[Unreleased]: https://github.com/giantswarm/app-operator/compare/v2.8.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app-operator/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/giantswarm/app-operator/compare/v2.8.0...v3.0.0
 [2.8.0]: https://github.com/giantswarm/app-operator/compare/v2.7.0...v2.8.0
 [2.7.0]: https://github.com/giantswarm/app-operator/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/giantswarm/app-operator/compare/v2.5.0...v2.6.0